[Refactoring] ファイル名の統一 Fixes #25
