+++ lib/main.dart
@@ -1,15 +1,13 @@
 // 🐦 Flutter imports:
 import 'package:flutter/material.dart';
-
 // 📦 Package imports:
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
 // 🌎 Project imports:
-import 'package:notion_wordbook/screens/connecting.dart';
-import 'package:notion_wordbook/screens/home.dart';
-import 'package:notion_wordbook/screens/test.dart';
-import 'package:notion_wordbook/screens/wordbook_item.dart';
+import 'package:notion_wordbook/screens/connecting_page.dart';
+import 'package:notion_wordbook/screens/home_page.dart';
+import 'package:notion_wordbook/screens/test_page.dart';
+import 'package:notion_wordbook/screens/wordbook_item_page.dart';
 import 'package:notion_wordbook/widgets/bottom_navbar.dart';
 
 void main() async {
