+++ lib/objects/models/word.dart
@@ -1,13 +1,13 @@
 // 🌎 Project imports:
 import 'package:notion_wordbook/objects/enums/word_tag.dart';
 
-class WordModel {
+class Word {
   final String spelling;
   final String? meaning, exampleSentence, link;
   // 正誤判定はAPIのJSON構造から必ず存在する
   final bool correct;
   final List<WordTag>? tags;
-  const WordModel(
+  const Word(
     this.spelling,
     this.correct,
     this.meaning,
